include imagery for multi-selection
